Divide alert into submodules
